Feature/limosek/dpkg status

+++ debian/letheand-fetchbc.sh
@@ -2,7 +2,7 @@
 
 . /etc/default/lethean-daemon
 
-if [ "$(whoami)" != "lthn" ]; then
+if [ "$(whoami)" = "root" ]; then
   echo "This script must be run as lthn user."
   exit 3
 fi
@@ -13,24 +13,29 @@ if [ -z "${LETHEAND_LMDB}" ] || [ -z "${ZSYNC_URL}" ]; then
   exit
 fi
 
+echo $$ >${LETHEAND_PID}
+
 download(){
+  if [ -n "$(pidof zsync)" ]; then
+    letheand status
+    exit 2
+  fi
   if ! fuser -s ${LETHEAND_LMDB}/data.mdb; then
-    echo "Downloading blockchain data. It can take long time. Be patient.." >&2
+    echo "Downloading blockchain data from ${ZSYNC_URL}. It can take long time. Be patient.." >&2
     mkdir -p ${LETHEAND_LMDB} \
     && cd ${LETHEAND_LMDB} \
     && rm -f data.mdb.zsync && \
-    wget "$ZSYNC_URL" && zsync data.mdb.zsync;
+    wget "$ZSYNC_URL" && zsync data.mdb.zsync && rm -f ${LETHEAND_PID}
   else
     echo "${LETHEAND_LMDB}/data.mdb in use by "$(fuser -v ${LETHEAND_LMDB}/data.mdb) '! Not downloading.' >&2
   fi
 }
 
 if [ -f ${LETHEAND_LMDB}/data.mdb ] && [ "$1" = "force" ] ; then \
-  echo "Downloading new blockchain data from ${ZSYNC_URL}"
+  shift
   download
 else
   if ! [ -f ${LETHEAND_LMDB}/data.mdb ]; then
-    echo "Downloading blockchain data from ${ZSYNC_URL}" >&2
     download
   else
     echo "Not touching blockchain data. use $0 force to redownload." >&2
